get all works & change create work fields

+++ src/controllers/work-controller.js
@@ -14,23 +14,33 @@ exports.createWork = catchAsync(async (req, res, next) => {
 
   const workData = { ...req.body, photographerId: req.user.id };
 
+  const mainImgUpload = await uploadService.upload(req.files.mainImage[0].path);
+  workData.mainImageUrl = mainImgUpload.secure_url;
+  workData.mainImageKey = mainImgUpload.public_id;
+  fs.unlink(req.files.mainImage[0].path);
+
   const newWork = await workService.createWork(workData);
 
   const imgData = [];
-  for (const file of req.files) {
-    const result = await uploadService.upload(file.path);
-    imgData.push({
-      imageName: file.filename.split(".")[0],
-      imageUrl: result,
-      workId: newWork.id,
-    });
-    fs.unlink(file.path);
-  }
 
+  if (req.files.subImages) {
+    for (const file of req.files.subImages) {
+      const result = await uploadService.upload(file.path);
+      imgData.push({
+        imageUrl: result.secure_url,
+        publicKey: result.public_id,
+        workId: newWork.id,
+      });
+      fs.unlink(file.path);
+    }
+  }
   await workImageService.createWorkImage(imgData);
 
   const work = await workService.findWorkAndWorkImageByWorkId(newWork.id);
   res.status(201).json({ work });
 });
 
-exports.getAllWorks = catchAsync(async (req, res, next) => {});
+exports.getAllWorks = catchAsync(async (req, res, next) => {
+  const works = await workService.findAllWorks();
+  res.status(200).json({ works });
+});
